+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `valid_rx_headers` method
 - `valid_rx_packets` method
 - `last_packet_snr` method
+- `set_invert_iq` method
 
 ## [0.1.0] - 2026-03-07